fixed some problems on Linux (GCC, Clang)

+++ libraries/include/ogg/os_types.h
@@ -141,13 +141,13 @@
 #elif defined(__ANDROID__) || defined(__linux__)
 
    /* Android GCC */
-#  include <sys/types.h>
-   typedef short ogg_int16_t;
-   typedef unsigned short ogg_uint16_t;
-   typedef int ogg_int32_t;
-   typedef unsigned int ogg_uint32_t;
-   typedef long long ogg_int64_t;
-   typedef unsigned long long ogg_uint64_t;
+#  include <stdint.h>
+   typedef int16_t ogg_int16_t;
+   typedef uint16_t ogg_uint16_t;
+   typedef int32_t ogg_int32_t;
+   typedef uint32_t ogg_uint32_t;
+   typedef int64_t ogg_int64_t;
+   typedef uint64_t ogg_uint64_t;
 
 #else
